+++ ir_sim/usage/19collision_mode/collision_mode.py
@@ -1,6 +1,6 @@
 from ir_sim.env import EnvBase
 
-env = EnvBase('collision_mode.yaml', control_mode='keyboard', collision_mode='react')
+env = EnvBase('collision_mode.yaml', control_mode='keyboard', collision_mode='unobstructed')
 # env = EnvBase('collision_mode_car.yaml', control_mode='keyboard', collision_mode='stop')
 
 for i in range(300):