+++ changelog.md
@@ -1,4 +1,12 @@
 
+## ir_sim 1.1.10
+
+- Fix some bugs
+- Add the collision mode 'unobstructed'
+- Add draw box function
+- Add radius for obstacle
+
+
 ## ir_sim 1.1.9
 
 - Fix some bugs